Get the PRIO-GRID bibliography

The bibliography contains all citations and further references that are included in pgsources. You can also get a subset of references, and you can get the RefManageR::RefManageR-package bibliography, or references as biblatex.

Usage

get_bibliography(keys = NULL, as_biblatex = FALSE)

Arguments

keys

A vector of strings, bibkeys found in pgsources.

as_biblatex

Set to true if you want results in biblatex instead of an R object.

Value

BibEntry or BibTex

Examples

get_bibliography(keys = "schvitzMappingInternationalSystem2022")
#> [1] G. Schvitz, L. Girardin, S. Rüegger, et al. “Mapping the
#> International System, 1886-2019: The CShapes 2.0 Dataset”. In: _Journal
#> of Conflict Resolution_ 66.1 (Jan. 2022), pp. 144-161. ISSN: 0022-0027.
#> DOI: 10.1177/00220027211013563. (Visited on 11/22/2024).