Get a sequence of date intervals

The start date is the last observed date in the first interval of the intervals you want to capture. E.g., if start date is 31 January 2010 with 1 month resolution, then the first interval goes from 1 January 2010 to 31 January 2010. The intervals are not necessary equal length.

Usage

pg_date_intervals(
  start_date = pgoptions$get_start_date(),
  end_date = pgoptions$get_end_date(),
  temporal_resolution = pgoptions$get_temporal_resolution()
)

Arguments

from

starting date

to

end date

by

increment of sequence. See details in base::seq.Date.

Value

Date interval vector

Examples

pg_date_intervals()
#>   [1] 1850-01-01 UTC--1850-12-31 UTC 1851-01-01 UTC--1851-12-31 UTC
#>   [3] 1852-01-01 UTC--1852-12-31 UTC 1853-01-01 UTC--1853-12-31 UTC
#>   [5] 1854-01-01 UTC--1854-12-31 UTC 1855-01-01 UTC--1855-12-31 UTC
#>   [7] 1856-01-01 UTC--1856-12-31 UTC 1857-01-01 UTC--1857-12-31 UTC
#>   [9] 1858-01-01 UTC--1858-12-31 UTC 1859-01-01 UTC--1859-12-31 UTC
#>  [11] 1860-01-01 UTC--1860-12-31 UTC 1861-01-01 UTC--1861-12-31 UTC
#>  [13] 1862-01-01 UTC--1862-12-31 UTC 1863-01-01 UTC--1863-12-31 UTC
#>  [15] 1864-01-01 UTC--1864-12-31 UTC 1865-01-01 UTC--1865-12-31 UTC
#>  [17] 1866-01-01 UTC--1866-12-31 UTC 1867-01-01 UTC--1867-12-31 UTC
#>  [19] 1868-01-01 UTC--1868-12-31 UTC 1869-01-01 UTC--1869-12-31 UTC
#>  [21] 1870-01-01 UTC--1870-12-31 UTC 1871-01-01 UTC--1871-12-31 UTC
#>  [23] 1872-01-01 UTC--1872-12-31 UTC 1873-01-01 UTC--1873-12-31 UTC
#>  [25] 1874-01-01 UTC--1874-12-31 UTC 1875-01-01 UTC--1875-12-31 UTC
#>  [27] 1876-01-01 UTC--1876-12-31 UTC 1877-01-01 UTC--1877-12-31 UTC
#>  [29] 1878-01-01 UTC--1878-12-31 UTC 1879-01-01 UTC--1879-12-31 UTC
#>  [31] 1880-01-01 UTC--1880-12-31 UTC 1881-01-01 UTC--1881-12-31 UTC
#>  [33] 1882-01-01 UTC--1882-12-31 UTC 1883-01-01 UTC--1883-12-31 UTC
#>  [35] 1884-01-01 UTC--1884-12-31 UTC 1885-01-01 UTC--1885-12-31 UTC
#>  [37] 1886-01-01 UTC--1886-12-31 UTC 1887-01-01 UTC--1887-12-31 UTC
#>  [39] 1888-01-01 UTC--1888-12-31 UTC 1889-01-01 UTC--1889-12-31 UTC
#>  [41] 1890-01-01 UTC--1890-12-31 UTC 1891-01-01 UTC--1891-12-31 UTC
#>  [43] 1892-01-01 UTC--1892-12-31 UTC 1893-01-01 UTC--1893-12-31 UTC
#>  [45] 1894-01-01 UTC--1894-12-31 UTC 1895-01-01 UTC--1895-12-31 UTC
#>  [47] 1896-01-01 UTC--1896-12-31 UTC 1897-01-01 UTC--1897-12-31 UTC
#>  [49] 1898-01-01 UTC--1898-12-31 UTC 1899-01-01 UTC--1899-12-31 UTC
#>  [51] 1900-01-01 UTC--1900-12-31 UTC 1901-01-01 UTC--1901-12-31 UTC
#>  [53] 1902-01-01 UTC--1902-12-31 UTC 1903-01-01 UTC--1903-12-31 UTC
#>  [55] 1904-01-01 UTC--1904-12-31 UTC 1905-01-01 UTC--1905-12-31 UTC
#>  [57] 1906-01-01 UTC--1906-12-31 UTC 1907-01-01 UTC--1907-12-31 UTC
#>  [59] 1908-01-01 UTC--1908-12-31 UTC 1909-01-01 UTC--1909-12-31 UTC
#>  [61] 1910-01-01 UTC--1910-12-31 UTC 1911-01-01 UTC--1911-12-31 UTC
#>  [63] 1912-01-01 UTC--1912-12-31 UTC 1913-01-01 UTC--1913-12-31 UTC
#>  [65] 1914-01-01 UTC--1914-12-31 UTC 1915-01-01 UTC--1915-12-31 UTC
#>  [67] 1916-01-01 UTC--1916-12-31 UTC 1917-01-01 UTC--1917-12-31 UTC
#>  [69] 1918-01-01 UTC--1918-12-31 UTC 1919-01-01 UTC--1919-12-31 UTC
#>  [71] 1920-01-01 UTC--1920-12-31 UTC 1921-01-01 UTC--1921-12-31 UTC
#>  [73] 1922-01-01 UTC--1922-12-31 UTC 1923-01-01 UTC--1923-12-31 UTC
#>  [75] 1924-01-01 UTC--1924-12-31 UTC 1925-01-01 UTC--1925-12-31 UTC
#>  [77] 1926-01-01 UTC--1926-12-31 UTC 1927-01-01 UTC--1927-12-31 UTC
#>  [79] 1928-01-01 UTC--1928-12-31 UTC 1929-01-01 UTC--1929-12-31 UTC
#>  [81] 1930-01-01 UTC--1930-12-31 UTC 1931-01-01 UTC--1931-12-31 UTC
#>  [83] 1932-01-01 UTC--1932-12-31 UTC 1933-01-01 UTC--1933-12-31 UTC
#>  [85] 1934-01-01 UTC--1934-12-31 UTC 1935-01-01 UTC--1935-12-31 UTC
#>  [87] 1936-01-01 UTC--1936-12-31 UTC 1937-01-01 UTC--1937-12-31 UTC
#>  [89] 1938-01-01 UTC--1938-12-31 UTC 1939-01-01 UTC--1939-12-31 UTC
#>  [91] 1940-01-01 UTC--1940-12-31 UTC 1941-01-01 UTC--1941-12-31 UTC
#>  [93] 1942-01-01 UTC--1942-12-31 UTC 1943-01-01 UTC--1943-12-31 UTC
#>  [95] 1944-01-01 UTC--1944-12-31 UTC 1945-01-01 UTC--1945-12-31 UTC
#>  [97] 1946-01-01 UTC--1946-12-31 UTC 1947-01-01 UTC--1947-12-31 UTC
#>  [99] 1948-01-01 UTC--1948-12-31 UTC 1949-01-01 UTC--1949-12-31 UTC
#> [101] 1950-01-01 UTC--1950-12-31 UTC 1951-01-01 UTC--1951-12-31 UTC
#> [103] 1952-01-01 UTC--1952-12-31 UTC 1953-01-01 UTC--1953-12-31 UTC
#> [105] 1954-01-01 UTC--1954-12-31 UTC 1955-01-01 UTC--1955-12-31 UTC
#> [107] 1956-01-01 UTC--1956-12-31 UTC 1957-01-01 UTC--1957-12-31 UTC
#> [109] 1958-01-01 UTC--1958-12-31 UTC 1959-01-01 UTC--1959-12-31 UTC
#> [111] 1960-01-01 UTC--1960-12-31 UTC 1961-01-01 UTC--1961-12-31 UTC
#> [113] 1962-01-01 UTC--1962-12-31 UTC 1963-01-01 UTC--1963-12-31 UTC
#> [115] 1964-01-01 UTC--1964-12-31 UTC 1965-01-01 UTC--1965-12-31 UTC
#> [117] 1966-01-01 UTC--1966-12-31 UTC 1967-01-01 UTC--1967-12-31 UTC
#> [119] 1968-01-01 UTC--1968-12-31 UTC 1969-01-01 UTC--1969-12-31 UTC
#> [121] 1970-01-01 UTC--1970-12-31 UTC 1971-01-01 UTC--1971-12-31 UTC
#> [123] 1972-01-01 UTC--1972-12-31 UTC 1973-01-01 UTC--1973-12-31 UTC
#> [125] 1974-01-01 UTC--1974-12-31 UTC 1975-01-01 UTC--1975-12-31 UTC
#> [127] 1976-01-01 UTC--1976-12-31 UTC 1977-01-01 UTC--1977-12-31 UTC
#> [129] 1978-01-01 UTC--1978-12-31 UTC 1979-01-01 UTC--1979-12-31 UTC
#> [131] 1980-01-01 UTC--1980-12-31 UTC 1981-01-01 UTC--1981-12-31 UTC
#> [133] 1982-01-01 UTC--1982-12-31 UTC 1983-01-01 UTC--1983-12-31 UTC
#> [135] 1984-01-01 UTC--1984-12-31 UTC 1985-01-01 UTC--1985-12-31 UTC
#> [137] 1986-01-01 UTC--1986-12-31 UTC 1987-01-01 UTC--1987-12-31 UTC
#> [139] 1988-01-01 UTC--1988-12-31 UTC 1989-01-01 UTC--1989-12-31 UTC
#> [141] 1990-01-01 UTC--1990-12-31 UTC 1991-01-01 UTC--1991-12-31 UTC
#> [143] 1992-01-01 UTC--1992-12-31 UTC 1993-01-01 UTC--1993-12-31 UTC
#> [145] 1994-01-01 UTC--1994-12-31 UTC 1995-01-01 UTC--1995-12-31 UTC
#> [147] 1996-01-01 UTC--1996-12-31 UTC 1997-01-01 UTC--1997-12-31 UTC
#> [149] 1998-01-01 UTC--1998-12-31 UTC 1999-01-01 UTC--1999-12-31 UTC
#> [151] 2000-01-01 UTC--2000-12-31 UTC 2001-01-01 UTC--2001-12-31 UTC
#> [153] 2002-01-01 UTC--2002-12-31 UTC 2003-01-01 UTC--2003-12-31 UTC
#> [155] 2004-01-01 UTC--2004-12-31 UTC 2005-01-01 UTC--2005-12-31 UTC
#> [157] 2006-01-01 UTC--2006-12-31 UTC 2007-01-01 UTC--2007-12-31 UTC
#> [159] 2008-01-01 UTC--2008-12-31 UTC 2009-01-01 UTC--2009-12-31 UTC
#> [161] 2010-01-01 UTC--2010-12-31 UTC 2011-01-01 UTC--2011-12-31 UTC
#> [163] 2012-01-01 UTC--2012-12-31 UTC 2013-01-01 UTC--2013-12-31 UTC
#> [165] 2014-01-01 UTC--2014-12-31 UTC 2015-01-01 UTC--2015-12-31 UTC
#> [167] 2016-01-01 UTC--2016-12-31 UTC 2017-01-01 UTC--2017-12-31 UTC
#> [169] 2018-01-01 UTC--2018-12-31 UTC 2019-01-01 UTC--2019-12-31 UTC
#> [171] 2020-01-01 UTC--2020-12-31 UTC 2021-01-01 UTC--2021-12-31 UTC
#> [173] 2022-01-01 UTC--2022-12-31 UTC 2023-01-01 UTC--2023-12-31 UTC
#> [175] 2024-01-01 UTC--2024-12-31 UTC